Fix vector gen, always enable BlockMap tests
The BlockMap tests got much faster, so there's no need to hide them
behind a flag.

+++ worldedit-core/src/test/java/com/sk89q/worldedit/util/test/VariedVectorGenerator.java
@@ -40,9 +40,9 @@ public class VariedVectorGenerator {
     public final int divisionsXZ;
     public final int divisionsY;
     public final int maxXZ;
-    public final int maxY;
+    public final long maxY;
     public final int xzStep;
-    public final int yStep;
+    public final long yStep;
     public final Set<BlockVector3> alwaysInclude;
 
     public VariedVectorGenerator() {
